+++ src/resmom/mom_hook_func.c
@@ -4071,130 +4071,6 @@ mom_process_hooks(unsigned int hook_event, char *req_user, char *req_host,
 	return (1);
 }
 
-
-/**
- * @brief
- *
- *	Returns the number of hook scripts that are eligible to
- *	be executed for the specified 'hook_event'.
- *	This means the hook is enabled and has a hook content.
- *
- * @param[in] 	hook_event - the event of the hooks to count
- * @return	int
- * @retval 	<n> number of hooks
- *
- */
-int
-num_eligible_hooks(unsigned int hook_event)
-{
-	hook			*phook;
-	hook			*phook_next = NULL;
-	pbs_list_head		*head_ptr;
-	int			num_hooks = 0;
-
-	switch (hook_event) {
-
-		case HOOK_EVENT_EXECJOB_BEGIN:
-			head_ptr = &svr_execjob_begin_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_PROLOGUE:
-			head_ptr = &svr_execjob_prologue_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_EPILOGUE:
-			head_ptr = &svr_execjob_epilogue_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_END:
-			head_ptr = &svr_execjob_end_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_PRETERM:
-			head_ptr = &svr_execjob_preterm_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_LAUNCH:
-			head_ptr = &svr_execjob_launch_hooks;
-			break;
-		case HOOK_EVENT_EXECHOST_PERIODIC:
-			head_ptr = &svr_exechost_periodic_hooks;
-			break;
-		case HOOK_EVENT_EXECHOST_STARTUP:
-			head_ptr = &svr_exechost_startup_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_ATTACH:
-			head_ptr = &svr_execjob_attach_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_RESIZE:
-			head_ptr = &svr_execjob_resize_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_ABORT:
-			head_ptr = &svr_execjob_abort_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_POSTSUSPEND:
-			head_ptr = &svr_execjob_postsuspend_hooks;
-			break;
-		case HOOK_EVENT_EXECJOB_PRERESUME:
-			head_ptr = &svr_execjob_preresume_hooks;
-			break;
-		default:
-			return (0); /* unexpected event encountered */
-	}
-
-	for (phook = (hook *)GET_NEXT(*head_ptr); phook; phook = phook_next) {
-		switch (hook_event) {
-
-			case HOOK_EVENT_EXECJOB_BEGIN:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_begin_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_PROLOGUE:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_prologue_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_EPILOGUE:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_epilogue_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_END:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_end_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_PRETERM:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_preterm_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_LAUNCH:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_launch_hooks);
-				break;
-			case HOOK_EVENT_EXECHOST_PERIODIC:
-				phook_next = (hook *)GET_NEXT(phook->hi_exechost_periodic_hooks);
-				break;
-			case HOOK_EVENT_EXECHOST_STARTUP:
-				phook_next = (hook *)GET_NEXT(phook->hi_exechost_startup_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_ATTACH:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_attach_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_RESIZE:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_resize_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_ABORT:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_abort_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_POSTSUSPEND:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_postsuspend_hooks);
-				break;
-			case HOOK_EVENT_EXECJOB_PRERESUME:
-				phook_next = (hook *)GET_NEXT(phook->hi_execjob_preresume_hooks);
-				break;
-			default:
-				return (0); /*  should not get here */
-		}
-
-		if (phook->enabled == FALSE)
-			continue;
-
-		if (phook->script == NULL)
-			continue;
-
-		num_hooks++;
-	}
-
-	return (num_hooks);
-}
-
 /**
  * @brief
  * 	Cleans up files older than HOOKS_TMPFILE_MAX_AGE under
