+++ src/server/req_jobobit.c
@@ -38,26 +38,8 @@
  */
 
 /**
- * @file    req_jobobit.c
- *
- * @brief
- * 		req_jobobit.c - functions dealing with a Job Obituary Request (Notice)
- *		and the associated post execution job clean up.
- *
- * Functions included are:
- * 	setup_from()
- * 	setup_cpyfiles()
- * 	is_joined()
- * 	cpy_stdfile()
- * 	cpy_stage()
- * 	mom_comm()
- * 	rel_resc()
- * 	on_exitrerun_msg()
- * 	conn_to_mom_failed()
- * 	on_job_exit()
- * 	on_job_rerun()
- * 	setrerun()
- * 	job_obit()
+ * functions dealing with a Job Obituary Request (Notice)
+ * and the associated post execution job clean up
  */
 
 #include <pbs_config.h>   /* the master config generated by configure */
@@ -132,392 +114,7 @@ void	on_job_rerun(struct work_task *ptask);
 void	on_job_exit(struct work_task *ptask);
 extern void set_admin_suspend(job *pjob, int set_remove_nstate);
 
-/* Local private functions */
-
-static struct batch_request *setup_cpyfiles
-(struct batch_request *, job *, char *,  char *, int, int);
-
 static char *msg_obitnotrun = "job not running, may have been requeued on node failure";
-/**
- * @brief
- * 		setup_from - setup the "from" name for a standard job file:
- *		output, error, or chkpt
- *
- * @param[in]	pjob	- job structure
- * @param[in]	suffix	- suffix for the "from" name
- *
- * @return	"from" name
- */
-
-static char *
-setup_from(job  *pjob, char *suffix)
-{
-	char *from;
-
-	from = malloc(strlen(pjob->ji_qs.ji_jobid) + strlen(suffix) + 1);
-	if (from) {
-		(void)strcpy(from, pjob->ji_qs.ji_jobid);
-		(void)strcat(from, suffix);
-	}
-	return (from);
-}
-
-/**
- * @brief
- * 		setup_cpyfiles - if need be, allocate and initialize a Copy Files
- *		batch request, then append the file pairs
- *
- * @param[in]	preq	- batch request
- * @param[in]	pjob	- job structure
- * @param[in]	from	- local (to mom) name
- * @param[in]	to		- remote (destination) name
- * @param[in]	direction	- copy direction
- * @param[in]	tflag	- 1 if stdout or stderr , 2 if stage out or in
- *
- * @return	modified batch request.
- * @retval	NULL	- failure
- */
-
-static
-struct batch_request *
-setup_cpyfiles(struct batch_request *preq, job  *pjob, char *from, char *to, int  direction, int  tflag)
-{
-	struct rq_cpyfile *pcf;
-	struct rq_cpyfile_cred *pcfc;
-	struct rqfpair    *pair;
-	size_t		  cred_len;
-	char		  *cred = NULL;
-	char		  *prq_jobid;
-	char		  *prq_owner;
-	char		  *prq_user;
-	char		  *prq_group;
-	int 		  *prq_dir;
-	pbs_list_head	  *prq_pair;
-
-	/* if this is a sub job of an array job, then check to see if the */
-	/* index needs to be substituted in the paths			  */
-
-	if (pjob->ji_qs.ji_svrflags & JOB_SVFLG_SubJob) {
-		to   = subst_array_index(pjob, to);
-		from = subst_array_index(pjob, from);
-	}
-
-	if (preq == NULL) {
-		char		*momname;
-
-		/* check that certain required attributues are valid */
-
-		if ((pjob->ji_wattr[(int)JOB_ATR_job_owner].at_val.at_str == NULL) ||
-			(pjob->ji_wattr[(int)JOB_ATR_euser].at_val.at_str == NULL)) {
-			/* this case shouldn't happen, log it and don't do copy     */
-			/* use null jobid, if attr missing, jobid is likely bad too */
-
-			log_event(PBSEVENT_ERROR|PBSEVENT_JOB, PBS_EVENTCLASS_FILE,
-				LOG_INFO,  "",
-				"cannot copy files for job, owner/euser missing");
-			if (from)
-				free(from);
-			if (to)
-				free(to);
-			if (cred)
-				free(cred);
-			return NULL;
-		}
-		/* allocate and initialize the batch request struct */
-
-		momname = parse_servername(pjob->ji_wattr[(int)JOB_ATR_exec_vnode].
-			at_val.at_str, NULL);
-		if (get_credential(momname, pjob, PBS_GC_CPYFILE,
-			&cred, &cred_len) == 0) {
-			preq = alloc_br(PBS_BATCH_CopyFiles_Cred);
-		} else {
-			preq = alloc_br(PBS_BATCH_CopyFiles);
-		}
-
-		if (preq == NULL) {
-			if (from)
-				free(from);
-			if (to)
-				free(to);
-			if (cred)
-				free(cred);
-			return (preq);
-		}
-
-		if (preq->rq_type == PBS_BATCH_CopyFiles_Cred) {
-			preq->rq_ind.rq_cpyfile_cred.rq_credtype =
-				pjob->ji_extended.ji_ext.ji_credtype;
-			preq->rq_ind.rq_cpyfile_cred.rq_pcred = cred;
-			preq->rq_ind.rq_cpyfile_cred.rq_credlen = cred_len;
-			pcfc = &preq->rq_ind.rq_cpyfile_cred;
-			prq_jobid = pcfc->rq_copyfile.rq_jobid;
-			prq_owner = pcfc->rq_copyfile.rq_owner;
-			prq_user = pcfc->rq_copyfile.rq_user;
-			prq_group = pcfc->rq_copyfile.rq_group;
-			prq_dir = &pcfc->rq_copyfile.rq_dir;
-			prq_pair = &pcfc->rq_copyfile.rq_pair;
-		} else {
-			pcf = &preq->rq_ind.rq_cpyfile;
-			prq_jobid = pcf->rq_jobid;
-			prq_owner = pcf->rq_owner;
-			prq_user = pcf->rq_user;
-			prq_group = pcf->rq_group;
-			prq_dir = &pcf->rq_dir;
-			prq_pair = &pcf->rq_pair;
-		}
-		CLEAR_HEAD((*prq_pair));
-
-		/* copy jobid, owner, exec-user, group names, upto the @host part */
-
-		(void)strcpy(prq_jobid, pjob->ji_qs.ji_jobid);
-		get_jobowner(pjob->ji_wattr[(int)JOB_ATR_job_owner].at_val.at_str,
-			prq_owner);
-		get_jobowner(pjob->ji_wattr[(int)JOB_ATR_euser].at_val.at_str,
-			prq_user);
-		if (((pjob->ji_wattr[(int)JOB_ATR_egroup].at_flags &
-			ATR_VFLAG_DEFLT) ==0) &&
-			(pjob->ji_wattr[(int)JOB_ATR_egroup].at_val.at_str != 0)) {
-			strcpy(prq_group,
-				pjob->ji_wattr[(int)JOB_ATR_egroup].at_val.at_str);
-		}
-		else
-			prq_group[0] = '\0';	/* default: use login group */
-
-		*prq_dir = direction;
-
-		/* set "sandbox=PRIVATE" mode */
-		if (pjob->ji_wattr[(int)JOB_ATR_sandbox].at_flags & ATR_VFLAG_SET) {
-			/* set STAGE_JOBDIR mode based on job settings */
-			if (strcasecmp(pjob->ji_wattr[(int)JOB_ATR_sandbox].at_val.at_str, "PRIVATE") == 0) {
-				*prq_dir |= STAGE_JOBDIR;
-			}
-		} /* O_WORKDIR check would go here */
-
-	} else {
-
-		/* use the existing request structure */
-
-		if (preq->rq_type == PBS_BATCH_CopyFiles_Cred) {
-			pcfc = &preq->rq_ind.rq_cpyfile_cred;
-			prq_pair = &pcfc->rq_copyfile.rq_pair;
-		} else {
-			pcf = &preq->rq_ind.rq_cpyfile;
-			prq_pair = &pcf->rq_pair;
-		}
-	}
-
-	pair = (struct rqfpair *)malloc(sizeof(struct rqfpair));
-	if (pair == NULL) {
-		free(from);
-		free(to);
-		free_br(preq);
-		return NULL;
-	}
-
-	CLEAR_LINK(pair->fp_link);
-	pair->fp_local  = from;
-	pair->fp_rmt    = to;
-	pair->fp_flag   = tflag;
-	append_link(prq_pair, &pair->fp_link, pair);
-	return (preq);
-}
-/**
- * @brief
- * 		is_joined - Is the file joined to another.
- *
- * @param[in]	pjob	- job structure
- * @param[in]	ati	- job attribute, output/error path.
- *
- * @return	joined or not
- * @retval	0	- either the first or not in list.
- * @retval	1	- being joined.
- */
-static int
-is_joined(pjob, ati)
-job		*pjob;
-enum job_atr	 ati;
-{
-	char       key;
-	attribute *pattr;
-	char	  *pd;
-
-	if (ati == JOB_ATR_outpath)
-		key = 'o';
-	else if (ati == JOB_ATR_errpath)
-		key = 'e';
-	else
-		return (0);
-	pattr = &pjob->ji_wattr[(int)JOB_ATR_join];
-	if (pattr->at_flags & ATR_VFLAG_SET) {
-		pd = pattr->at_val.at_str;
-		if (pd && *pd && (*pd != 'n')) {
-			/* if not the first letter, and in list - is joined */
-			if ((*pd != key) && (strchr(pd+1, (int)key)))
-				return (1);	/* being joined */
-		}
-	}
-	return (0);	/* either the first or not in list */
-}
-
-/**
- * @brief
- * 		cpy_stdfile - determine if one of the job's standard files (output or error)
- *		is to be copied, if so set up the Copy Files request.
- *
- * @param[in]	preq	- batch request
- * @param[in]	pjob	- job structure
- * @param[in]	ati	- JOB_ATR_, output/error path.
- *
- * @return	modified batch request.
- * @retval	NULL	failure
- */
-
-static struct batch_request *cpy_stdfile(struct batch_request *preq, job *pjob, enum job_atr ati)
-{
-	char *from;
-	char  key;
-	attribute *jkpattr;
-	attribute *pathattr = &pjob->ji_wattr[(int)ati];
-	char *suffix;
-	char *to = NULL;
-
-	/* if the job is interactive, don't bother to return output file */
-
-	if (pjob->ji_wattr[(int)JOB_ATR_interactive].at_flags &&
-		pjob->ji_wattr[(int)JOB_ATR_interactive].at_val.at_long)
-		return NULL;
-
-	/* set up depending on which file */
-
-	if (ati == JOB_ATR_errpath) {
-		key    = 'e';
-		suffix = JOB_STDERR_SUFFIX;
-	} else {
-		key    = 'o';
-		suffix = JOB_STDOUT_SUFFIX;
-	}
-
-	if ((pathattr->at_flags & ATR_VFLAG_SET) == 0) { /* This shouldn't be */
-
-		(void)sprintf(log_buffer, "%c file missing", key);
-		log_event(PBSEVENT_ERROR|PBSEVENT_JOB, PBS_EVENTCLASS_JOB,
-			LOG_INFO,  pjob->ji_qs.ji_jobid, log_buffer);
-		return NULL;
-	}
-
-	/* Is the file joined to another, if so don't copy it */
-
-	if (is_joined(pjob, ati))
-		return (preq);
-
-	/*
-	 * If the job has a keep file attribute, and the specified file is in
-	 * the keep list, MOM has already placed the file in the user's HOME
-	 * directory.  It don't need to be copied.
-	 */
-
-	jkpattr = &pjob->ji_wattr[(int)JOB_ATR_keep];
-	if ((jkpattr->at_flags & ATR_VFLAG_SET) &&
-		strchr(jkpattr->at_val.at_str, key) && !strchr(jkpattr->at_val.at_str, 'd'))
-
-		return (preq);
-
-	/*
-	 * If the job has a remove file attribute and the job has succeeded,
-	 * std_files doesn't has to be copied.
-	 */
-	if ((pjob->ji_wattr[(int)JOB_ATR_exit_status].at_flags) &
-					ATR_VFLAG_SET) {
-		if (pjob->ji_wattr[(int) JOB_ATR_exit_status].at_val.at_long
-				== JOB_EXEC_OK) {
-			jkpattr = &pjob->ji_wattr[(int) JOB_ATR_remove];
-			if ((jkpattr->at_flags & ATR_VFLAG_SET)
-					&& (strchr(jkpattr->at_val.at_str, key)))
-				return (preq);
-		}
-	}
-
-	/* else go with the supplied name */
-
-	to = malloc(strlen(pathattr->at_val.at_str) + 1);
-	if (to) {
-		(void)strcpy(to, pathattr->at_val.at_str);
-
-	} else
-		return (preq);	/* cannot continue with this one */
-
-	/* build up the name used by MOM as the from name */
-
-	from = setup_from(pjob, suffix);
-	if (from == NULL) {
-		(void)free(to);
-		return (preq);
-	}
-
-	/* now set names into the batch request */
-
-	return (setup_cpyfiles(preq, pjob, from, to, STAGE_DIR_OUT, STDJOBFILE));
-}
-
-/**
- * @brief
- * 		cpy_stage - set up a Copy Files request to include files specified by the
- *		user to be staged out (also used for stage-in).
- *		"stage_out" is a resource that may or may not *	exist on a host.
- *		If such exists, the files are listed one per string as
- *		"local_name@remote_host:remote_name".
- *
- * @param[in]	preq	- batch request
- * @param[in]	pjob	- job structure
- * @param[in]	ati	- JOB_ATR_stageout
- * @param[in]	direction	-  1 = , 2 =
- *
- * @return	batch_request *
- */
-
-struct batch_request *cpy_stage(struct batch_request *preq, job *pjob, enum job_atr ati, int direction)
-{
-	int		      i;
-	char		     *from;
-	attribute 	     *pattr;
-	struct array_strings *parst;
-	char 		     *plocal;
-	char		     *prmt;
-	char		     *to;
-
-	pattr = &pjob->ji_wattr[(int)ati];
-	if (pattr->at_flags & ATR_VFLAG_SET) {
-
-		/* at last, we know we have files to stage out/in */
-
-		parst = pattr->at_val.at_arst;
-		for (i = 0; i<parst->as_usedptr; ++i) {
-			plocal = parst->as_string[i];
-			prmt   = strchr(plocal, (int)'@');
-			if (prmt) {
-				*prmt = '\0';
-				from = malloc(strlen(plocal)+1);
-				if (from) {
-					(void)strcpy(from, plocal);
-					*prmt = '@';	/* restore the @ */
-				} else {
-					return (preq);
-				}
-				to   = malloc(strlen(prmt+1) + 1);
-				if (to) {
-					(void)strcpy(to, prmt+1);
-				} else {
-					(void)free(from);
-					return (preq);
-				}
-				preq = setup_cpyfiles(preq, pjob, from, to,
-					direction, STAGEFILE);
-			}
-		}
-	}
-
-	return (preq);
-}
 
 
 /**
@@ -696,6 +293,60 @@ conn_to_mom_failed(job *pjob, void(*func)(struct work_task *))
 	return;
 }
 
+static void
+end_job(job *pjob, int isexpress)
+{
+	char *rec = "";
+
+	if (isexpress) {
+		log_event(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, LOG_DEBUG, pjob->ji_qs.ji_jobid, "express end of job");
+		/* see if have any dependencies */
+		if (pjob->ji_wattr[(int)JOB_ATR_depend].at_flags & ATR_VFLAG_SET)
+			(void)depend_on_term(pjob);
+
+		/* Set job's exec_vnodes with current time for last_used_time. */
+		set_last_used_time_node(pjob, 0);
+	}
+
+
+	if (pjob->ji_momhandle != -1 && pjob->ji_mom_prot == PROT_TCP)
+		svr_disconnect(pjob->ji_momhandle);
+	rel_resc(pjob); /* free any resc assigned to the job */
+
+	account_job_update(pjob, PBS_ACCT_LAST);
+	account_jobend(pjob, pjob->ji_acctrec, PBS_ACCT_END);
+
+	if (pjob->ji_acctrec)
+		rec = pjob->ji_acctrec;
+
+	if (server.sv_attr[SVR_ATR_log_events].at_val.at_long &
+		PBSEVENT_JOB_USAGE) {
+		/* log events set to record usage */
+		log_event(PBSEVENT_JOB_USAGE | PBSEVENT_JOB_USAGE,
+			PBS_EVENTCLASS_JOB, LOG_INFO,
+			pjob->ji_qs.ji_jobid, rec);
+	} else {
+		char *pc;
+
+		/* no usage in log, truncate messge */
+
+		if ((pc = strchr(rec, ' ')) != NULL)
+			*pc = '\0';
+		log_event(PBSEVENT_JOB, PBS_EVENTCLASS_JOB, LOG_INFO,
+			pjob->ji_qs.ji_jobid, rec);
+	}
+
+	if ((pjob->ji_qs.ji_svrflags & JOB_SVFLG_HERE) == 0)
+		issue_track(pjob);
+
+	if (pjob->ji_pmt_preq != NULL)
+		reply_preempt_jobs_request(PBSE_NONE, PREEMPT_METHOD_DELETE, pjob);
+	/*
+	 * Check if the history of the finished job can be saved or it needs to be purged.
+	 */
+	svr_saveorpurge_finjobhist(pjob);
+}
+
 /**
  * @brief
  * 		continue post-execution processing of a job that terminated.
@@ -722,7 +373,6 @@ on_job_exit(struct work_task *ptask)
 	job   *pjob;
 	struct batch_request *preq;
 	struct work_task *pt;
-	char  *rec;
 	int    rc;
 	int    stageout_status = 1; /* success */
 	long   t;
@@ -738,15 +388,30 @@ on_job_exit(struct work_task *ptask)
 		pjob = (job *)preq->rq_extra;
 	}
 
+		/* minor check on validity of pjob */
+	if (isdigit((int)pjob->ji_qs.ji_jobid[0]) == 0)
+		return;		/* not pointing to currently valid job */
+
+	if (pjob->ji_qs.ji_substate == JOB_SUBSTATE_EXITING) {
+		/*
+		 * If jobs doesn't have any files to stage/delete and there is no execjob_end
+		 * hook to run, end job immediately
+		 *
+		 * If no stage files but has execjob_end hook to run, put job directly in
+		 * exited sub state
+		 */
+		if ((rc = num_eligible_hooks(HOOK_EVENT_EXECJOB_END)) == 0 && !has_stage(pjob)) {
+			end_job(pjob, 1);
+			return;
+		} else if (rc > 0) {
+			svr_setjobstate(pjob, JOB_STATE_EXITING, JOB_SUBSTATE_EXITED);
+		}
+	}
+
 	if ((pjob->ji_wattr[(int)JOB_ATR_relnodes_on_stageout].at_flags & ATR_VFLAG_SET) &&
 	   (pjob->ji_wattr[(int)JOB_ATR_relnodes_on_stageout].at_val.at_long != 0)) {
 		release_nodes_on_stageout = 1;
 	}
-
-	/* minor check on validity of pjob */
-
-	if (isdigit((int)pjob->ji_qs.ji_jobid[0]) == 0)
-		return;		/* not pointing to currently valid job */
 
 	if ((handle = mom_comm(pjob, on_job_exit)) < 0)
 		return;
@@ -1041,44 +706,7 @@ on_job_exit(struct work_task *ptask)
 				/* all went ok with the delete by Mom(s) */
 				free_br(preq);
 				preq = NULL;
-				if (handle != -1 && pjob->ji_mom_prot == PROT_TCP)
-					svr_disconnect(handle);
-				rel_resc(pjob); /* free any resc assigned to the job */
-
-				account_job_update(pjob, PBS_ACCT_LAST);
-				account_jobend(pjob, pjob->ji_acctrec, PBS_ACCT_END);
-
-				if (pjob->ji_acctrec)
-					rec = pjob->ji_acctrec;
-				else
-					rec = "";
-
-				if (server.sv_attr[(int)SVR_ATR_log_events].at_val.at_long &
-					PBSEVENT_JOB_USAGE) {
-					/* log events set to record usage */
-					log_event(PBSEVENT_JOB_USAGE | PBSEVENT_JOB_USAGE,
-						PBS_EVENTCLASS_JOB, LOG_INFO,
-						pjob->ji_qs.ji_jobid, rec);
-				} else {
-					char *pc;
-
-					/* no usage in log, truncate messge */
-
-					if ((pc = strchr(rec, (int)' ')) != NULL)
-						*pc = '\0';
-					log_event(PBSEVENT_JOB, PBS_EVENTCLASS_JOB, LOG_INFO,
-						pjob->ji_qs.ji_jobid, rec);
-				}
-
-				if ((pjob->ji_qs.ji_svrflags & JOB_SVFLG_HERE) == 0)
-					issue_track(pjob);
-
-				if (pjob->ji_pmt_preq != NULL)
-					reply_preempt_jobs_request(PBSE_NONE, PREEMPT_METHOD_DELETE, pjob);
-				/*
-				 * Check if the history of the finished job can be saved or it needs to be purged.
-				 */
-				svr_saveorpurge_finjobhist(pjob);
+				end_job(pjob, 0);
 			}
 			break;
 		case JOB_SUBSTATE_TERMINATED:
